Ball Changed
Ball has been changed so on reset, when it bounces out of the screen it can head towards any of the 4 corners instead of just the lower right. It has also been changed so that on contact with a paddle, the angle and ball speed are mildly varied so it isn't always a 90 degree angle bounce. Appearance has been changed to a blue star instead of a rectangle.