Glitch effect
Added something that causes the Chiyos to flicker in and out like they're glitching. Not sure what it's doing exactly but presumably its just quickly redrawing the chiyos over and over again so they overlap and flicker. Also started trying to add a clickable target but that is not working yet.