Added sound
Added sound files. On a correct click, a meowing sound will play and on an incorrect click, a hissing sound will play. If the player wins, it'll play the sound of a cat purring and if the player loses, it's the sound of a cat crying. Previous commit was accidentally uploaded to the wrong folder...